requirements.txt
flask==3.0.0
python-dotenv==1.0.0
groq>=0.4.2
httpx<0.28.0
fpdf==1.7.2
